wip StreamerReceiverSerializer
