+++ modules/platforms/dotnet/Apache.Ignite/Internal/Table/Serialization/StreamerReceiverSerializer.cs
@@ -18,6 +18,7 @@
 namespace Apache.Ignite.Internal.Table.Serialization;
 
 using System;
+using System.Buffers;
 using System.Diagnostics;
 using Ignite.Table;
 using Proto.BinaryTuple;
@@ -79,5 +80,49 @@ internal static class StreamerReceiverSerializer
 
         w.Write(binaryTupleSize);
         w.Write(builder.Build().Span);
+    }
+
+    /// <summary>
+    /// Reads receiver execution results.
+    /// </summary>
+    /// <param name="reader">Reader.</param>
+    /// <typeparam name="T">Result element type.</typeparam>
+    /// <returns>Pooled array with results and the actual element count.</returns>
+    public static (T[]? ResultsPooledArray, int ResultsCount) ReadReceiverResults<T>(MsgPackReader reader)
+    {
+        if (reader.TryReadNil())
+        {
+            return (null, 0);
+        }
+
+        var numElements = reader.ReadInt32();
+        if (numElements == 0)
+        {
+            return (null, 0);
+        }
+
+        var tuple = new BinaryTupleReader(reader.ReadBinary(), numElements);
+        if (tuple.GetInt(0) != TupleWithSchemaMarshalling.TypeIdTuple)
+        {
+            return tuple.GetObjectCollectionWithType<T>();
+        }
+
+        int elementCount = tuple.GetInt(1);
+        T[] resultsPooledArr = ArrayPool<T>.Shared.Rent(elementCount);
+
+        try
+        {
+            for (var i = 0; i < elementCount; i++)
+            {
+                resultsPooledArr[i] = (T)(object)TupleWithSchemaMarshalling.Unpack(tuple.GetBytesSpan(2 + i));
+            }
+
+            return (resultsPooledArr, elementCount);
+        }
+        catch
+        {
+            ArrayPool<T>.Shared.Return(resultsPooledArr);
+            throw;
+        }
     }
 }
